requirements.txt
Flask-RESTful
flask_cors
Flask-APScheduler
requests
python-decouple
Flask-SQLAlchemy
beautifulsoup4
lxml